+++ tests/kevin_testing/src/changelog.txt
@@ -1,15 +1,15 @@
-Tokenizer
-
 Parser
 
-accepted_methods should be limited to GET POST DELETE (HEAD).
+Tokenizer
+
+Interpreter
+
 listen should be saved in a struct with address and port.
 error_page should be saved in a vector of structs with error_code and link.
 redirect should be a vector of structs with status_code, from & to.
 _parse_location_path error for wrong wildcard usage
 
-After-Parser
+Analyzer
 
-Need name for After-Parser.
-After-Parser needed for later
+Analzyer needed
 default server should be deleted and put to the back.
